// utils/types.ts
export interface UserFormData {
    lastName: string;
    firstName: string;
    phone: string;
    email: string;
    password: string;
  }
  
  export interface EditUserFormData {
    lastName: string;
    firstName: string;
    phone: string;
  }
  
  export interface PaginationDetails {
    from: number;
    to: number;
    total: number;
  }
  